tech-feed-back/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==1.5.3
python-dateutil==2.8.2
psycopg2-binary==2.9.9
kafka-python==2.0.2 